ramips: add support for TP-Link TL-WDR8610 v3
TP-Link TL-WDR8610 v3 is a dual-band AC2600 router, based on
MediaTek MT7621 + MT7615N + MT7611N. The router uses signed
firmware, so to install OpenWrt the factory u-boot must be
replaced with external programmer with another bootloader
that doesn't require validating firmware signatures - e.g
the RE650 v1 u-boot distributed with the RE650 v1 GPL sources.
This is pretty much a dead end for almost anyone wanting to
use OpenWrt. The device itself is otherwise capable of
running OpenWrt.

Hardware specification:

- SoC 880 MHz - MediaTek MT7621AT
- 128 MB of DDR3 RAM Nanya NT5CC64M16GP-DI
- 8 MB of FLASH - GigaDevice GD25Q64CSIG
- 2.4 GHz - MediaTek MT7615N
- 5 GHz - MediaTek MT7611N
- 5x 1 Gbps Ethernet - MT7621AT integrated
- 1 front panel LED (SYS), 5 rear panel LEDs (LAN + WAN)
- 1x button (Reset)
- UART header on PCB

Vendor firmware partitions:

0x000000-0x01d800 : "factory_boot"
0x01d800-0x01e000 : "factory_info" (MAC)
0x01e000-0x020000 : "art" (8KB, 5G data @0x1000)
0x020000-0x030000 : "config"
0x030000-0x040000 : "normal_boot"
0x040000-0x13fa78 : "kernel"
0x13fa78-0x680000 : "rootfs"
0x680000-0x800000 : "rootfs_data"
0x040000-0x800000 : "firmware"

Issues:

1) The TP-Link "slp" devices use a different firmware header.
Since the bootloader has to be replaced anyways, we don't
care of supporting that type of header. We can build an
u-boot from the TP-Link GPL sources of a similar MT7621
device that uses the v1 type of header. This patch moves
the firmware partition to offset 0x20000 which is what
the MT7621 u-boot in the source packages requires.

2) The MT7611N device is unknown MT7615 variant.
It works with the currect mt76 driver if you add the
PCI device id in the mt76.

Signed-off-by: Georgi Vlaev <georgi.vlaev@gmail.com>